kernel: repair terminal stop release lifecycle
Set ReleaseFailed and abort a Taken stop when relation retirement, space release, or CR3 restoration fails while keeping admission and frame accounting live. Finish every post-Taken terminal failure as Aborted, gate timer consumption on exact Running state, split manager trap/release responsibilities below the function cap, and split ktest event parsing while adding the observed stop register ordering.
